+++ Examples/Dynamic Controllers/Examples.DynamicControllers/Startup.cs
@@ -32,7 +32,8 @@ namespace Examples
         {
             templates
                 .AddChildrenControllerTemplates()
-                .AddParentsControllerTemplates();
+                .AddParentsControllerTemplates()
+                ;
 
             return templates;
         }